Extract TerminalGpuResources to resources.rs (refs #88)
Moves TerminalGpuResources struct + impl (retain_panes,
evict_unused_images) out of callback.rs into its own resources.rs module.
Mirrors wezterm-gui's separation of resource ownership from paint logic.

No behavior change — pure mechanical refactor.
