+++ crates/amux-render-gpu/src/shape.rs
@@ -12,8 +12,8 @@
 use cosmic_text::{Attrs, Buffer, Metrics, Shaping};
 use egui_wgpu::wgpu;
 
-use crate::callback::TerminalGpuResources;
 use crate::pipeline::CellFgInstance;
+use crate::resources::TerminalGpuResources;
 use crate::state::{CachedGlyph, ShapedGlyphEntry, TextRun};
 
 /// Shape a multi-cell text run for ligature support.
